Fix ptp2 camlib compilation warnings.
* chdk.c: Signed-unsigned comparison fixes.
* config.c: Signed-unsigned comparison and switch fall through fixes. Macro on line 9601 was modified to deal with signed/unsigned variables. Potentially loop still explodes if max range value happens to be max long/unsigned long value.
* library.c: Signed-unsigned comparison and switch fall through fixes.
* olympus-wrap.c: Switch fall through fix.
* ptp-pack.c: Signed-unsigned comparison fixes. Also starting from line 2171 there seems to be a bug that is now fixed. Previously, if payload is smaller than 0x0c bytes then 'PTP_EC_CANON_EOS_PropValueChanged' switch case will fall through.
* ptp.c: Signed-unsigned comparison fixes and remove one set of unnecessary braces.
* ptpip.c: Switch fall through fixes.
* usb.c: Signed-unsigned comparison fix.

+++ camlibs/ptp2/chdk.c
@@ -1005,7 +1005,7 @@ static struct {
 
 static int
 chdk_get_onoff(CONFIG_GET_ARGS) {
-        int i;
+        unsigned int i;
         char buf[1024];
 
         gp_widget_new (GP_WIDGET_RADIO, _(menu->label), widget);
@@ -1022,7 +1022,7 @@ chdk_get_onoff(CONFIG_GET_ARGS) {
 
 static int
 chdk_put_onoff(CONFIG_PUT_ARGS) {
-        int i;
+        unsigned int i;
         char *val;
 
         CR (gp_widget_get_value(widget, &val));
@@ -1214,7 +1214,7 @@ static void yuv_live_to_jpeg(unsigned char *p_yuv,
 		/* offset to the correct row */
 		unsigned int offset = cinfo.next_scanline * row_inc;
 
-		for (x = 0, i = 0, j = 0; x < width; i += sshift, j += dshift, x += xshift) {
+		for (x = 0, i = 0, j = 0; (int)x < width; i += sshift, j += dshift, x += xshift) {
 			int8_t u = (int8_t) p_yuv[offset + i + 0];
 			int8_t v = (int8_t) p_yuv[offset + i + 2];
 			if (fb_type == LV_FB_YUV8) {
